# swagger/oas30_gen.json
{"openapi":"3.0.0","info":{"title":"srv-podcasts","version":"1.0.0","description":"Podcast Service for AnyPodcast"},"paths":{"/podcast/{id}":{"get":{"operationId":"get","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"200 response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Podcast"}}}}},"x-amazon-apigateway-integration":{"httpMethod":"POST","uri":"arn:aws:apigateway:${region}:lambda:path/2015-03-31/functions/arn:aws:lambda:${region}:${account}:function:${lambda_name}/invocations","responses":{"default":{"statusCode":"200"}},"passthroughBehavior":"when_no_match","contentHandling":"CONVERT_TO_TEXT","type":"aws_proxy"}},"put":{"operationId":"update","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Podcast"}}},"required":true},"responses":{"200":{"description":"200 response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Empty"}}}}},"x-amazon-apigateway-integration":{"httpMethod":"POST","uri":"arn:aws:apigateway:${region}:lambda:path/2015-03-31/functions/arn:aws:lambda:${region}:${account}:function:${lambda_name}/invocations","responses":{"default":{"statusCode":"200"}},"passthroughBehavior":"when_no_match","contentHandling":"CONVERT_TO_TEXT","type":"aws_proxy"}},"delete":{"operationId":"delete","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"200 response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Empty"}}}}},"x-amazon-apigateway-integration":{"httpMethod":"POST","uri":"arn:aws:apigateway:${region}:lambda:path/2015-03-31/functions/arn:aws:lambda:${region}:${account}:function:${lambda_name}/invocations","responses":{"default":{"statusCode":"200"}},"passthroughBehavior":"when_no_match","contentHandling":"CONVERT_TO_TEXT","type":"aws_proxy"}}},"/podcast":{"get":{"operationId":"get_all","responses":{"200":{"description":"200 response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PodcastList"}}}}},"parameters":[{"name":"UserID","in":"query","required":true,"schema":{"type":"string"}},{"name":"SearchValue","in":"query","required":false,"schema":{"type":"string"}}],"x-amazon-apigateway-request-validator":"Validate query string parameters and headers","x-amazon-apigateway-integration":{"httpMethod":"POST","uri":"arn:aws:apigateway:${region}:lambda:path/2015-03-31/functions/arn:aws:lambda:${region}:${account}:function:${lambda_name}/invocations","responses":{"default":{"statusCode":"200"}},"passthroughBehavior":"when_no_match","contentHandling":"CONVERT_TO_TEXT","type":"aws_proxy"}},"post":{"operationId":"create","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Podcast"}}},"required":true},"responses":{"200":{"description":"200 response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Empty"}}}}},"x-amazon-apigateway-request-validator":"Validate body","x-amazon-apigateway-integration":{"httpMethod":"POST","uri":"arn:aws:apigateway:${region}:lambda:path/2015-03-31/functions/arn:aws:lambda:${region}:${account}:function:${lambda_name}/invocations","responses":{"default":{"statusCode":"200"}},"passthroughBehavior":"when_no_match","contentHandling":"CONVERT_TO_TEXT","type":"aws_proxy"}}}},"components":{"schemas":{"Empty":{"title":"Empty Schema","type":"object"},"Podcast":{"type":"object","properties":{"Created":{"type":"string","default":""},"UserID":{"type":"string","default":""},"Title":{"type":"string","default":""},"Description":{"type":"string","default":""},"IsVisible":{"type":"boolean","default":false}},"required":["Created","Description","IsVisible","Title","UserID"],"title":"Podcast"},"PodcastList":{"type":"array","items":{"$ref":"#/components/schemas/Podcast"},"title":"PodcastList"},"SearchParams":{"type":"object","properties":{"UserID":{"type":"string"},"SearchValue":{"type":"string"}},"required":["UserID"],"title":"SearchParams"}}},"tags":[],"servers":[],"x-amazon-apigateway-request-validators":{"Validate body":{"validateRequestParameters":false,"validateRequestBody":true},"Validate query string parameters and headers":{"validateRequestParameters":true,"validateRequestBody":false}}}